Move SessionPlanPoll to commands/plansession
